fix: useAuth hookで認証チェックを実現
